Fix: all warning and error related to theme provider
