+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import AppRoutes from './routes';
-import { ThemeProvider } from '@/components/ThemeProvider';
+import { ThemeProvider } from '@/components/ThemeProvider/ThemeProvider';
 
 function App() {
 	return (
