+++ frontend/src/components/ThemeProvider/useTheme.tsx
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import ThemeContext from "./ThemeContext";
+
+export function useTheme() {
+	const context = useContext(ThemeContext);
+	if (!context) {
+		throw new Error("useTheme must be used within a ThemeProvider");
+	}
+	return context;
+}
